Add 'loc' directive
Organisations can be "located at" other "host" organisations for the purposes
of contactability, in lieu of having their own address and phone numbers.

TODO: "Located in" relations are not shown in --report=book.

TODO: "Located in" relations are shown in --report=dump but but should not list
all the workers or departments of the host org, and should list all the parent
orgs.

TODO: "Located in" relations are shown in --report=phone, but should probably
not list all phone numbers of all the workers or departments of the host org,
just the org's own numbers and of its parent orgs.